+++ resdk/analysis/_register.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from resdk.analysis.alignment import bowtie2, hisat2
 from resdk.analysis.chip_seq import macs, rose2
+from resdk.analysis.differential_expressions import cuffdiff
 from resdk.analysis.expressions import cuffnorm, cuffquant
 from resdk.analysis.plots import bamliquidator, bamplot
 from resdk.resources import Collection, Relation, Sample
@@ -10,6 +11,7 @@ from resdk.resources import Collection, Relation, Sample
 Collection.run_bamliquidator = bamliquidator
 Collection.run_bamplot = bamplot
 Collection.run_bowtie2 = bowtie2
+Collection.run_cuffdiff = cuffdiff
 Collection.run_cuffnorm = cuffnorm
 Collection.run_cuffquant = cuffquant
 Collection.run_hisat2 = hisat2
@@ -19,6 +21,7 @@ Collection.run_rose2 = rose2
 Relation.run_bamliquidator = bamliquidator
 Relation.run_bamplot = bamplot
 Relation.run_bowtie2 = bowtie2
+Relation.run_cuffdiff = cuffdiff
 Relation.run_cuffnorm = cuffnorm
 Relation.run_cuffquant = cuffquant
 Relation.run_hisat2 = hisat2